# More accurate number parsing