We now achieve an [ULP](https://en.wikipedia.org/wiki/Unit_in_the_last_place) bounded by 1 (as an integer value) during number parsing. It should be as good as needed for most applications.

This harms slightly our performance on documents with many numbers.... 

Here are some speeds...
 
|   |  before | this PR   |
|---|---|---|
| canada  |  1.03 GB/s  | 0.94 GB/s   |
|  mesh | 0.91 GB/s  |  0.88 GB/s  |
|   mesh.pretty |  1.40 GB/s | 1.57 GB/s  |
|  number |  1.06 GB/s |  0.98 GB/s |

It is pretty much a small loss of performance *except* for canada and numbers where we drop by almost 10%. And mostly, canada is an adversarial case because it has many numbers like 74.133880999999974 and 118.82333399999999. Our parser tries to represent these values as integers that are later scaled back to floating-point numbers. However, they are too large to fit (exceeds 1<<53) so we suffer due to an expensive fallback.

Yet these two files (number and canada) are mostly "number parsing" cases... there are other ways to improve the performance over these files that we are going to implement later. Performance should go back up.

Once solved, this would solve https://github.com/lemire/simdjson/issues/71.