@lemire 2019-07-16
@ioioioio You should check the ULP computation as it could prove useful for testing!