@lemire 2019-07-16
This passes all tests, and the performance regression is limited to just some cases that we will revisit before the next release.